feat: implement more endpoints
