+++ crates/alerion_core/src/webserver.rs
@@ -1,7 +1,6 @@
 #![deny(dead_code)]
 use std::collections::HashSet;
 use std::env::consts::{ARCH, OS};
-use std::net::SocketAddr;
 
 use bitflags::bitflags;
 use futures::{SinkExt, StreamExt};
@@ -9,15 +8,14 @@ use jsonwebtoken::{Algorithm, DecodingKey, Validation};
 use poem::listener::TcpListener;
 use poem::middleware::Cors;
 use poem::web::websocket::{Message, WebSocket};
-use poem::web::Path;
-use poem::{
-    get, handler, Body, Endpoint, EndpointExt, IntoResponse, Middleware, Request, Response, Route, Server
-};
-use reqwest::{Method, StatusCode};
+use poem::web::{Json, Path};
+use poem::{endpoint, get, handler, EndpointExt, IntoResponse, Route, Server};
+use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 use sysinfo::System;
 use uuid::Uuid;
 
+use self::middleware::bearer_auth::BearerAuthMiddleware;
 use crate::config::AlerionConfig;
 
 //const ALLOWED_HEADERS: &str = "Accept, Accept-Encoding, Authorization, Cache-Control, Content-Type, Content-Length, Origin, X-Real-IP, X-CSRF-Token";
@@ -146,11 +144,6 @@ impl Auth {
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-struct SystemQuery {
-    v: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
 struct SystemResponseV1 {
     architecture: String,
     cpu_count: usize,
@@ -235,24 +228,17 @@ pub enum ServerEventType {
 #[handler]
 async fn process_system_query() -> impl IntoResponse {
     let Some(kernel_version) = System::kernel_version() else {
-        return Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .body("The system information could not be fetched.");
+        return StatusCode::INTERNAL_SERVER_ERROR.into_response();
     };
 
-    let Ok(response) = Body::from_json(SystemResponseV1 {
+    Json(SystemResponseV1 {
         architecture: ARCH.to_owned(),
         cpu_count: num_cpus::get(),
         kernel_version,
         os: OS.to_owned(),
         version: env!("CARGO_PKG_VERSION").to_owned(),
-    }) else {
-        return Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .finish();
-    };
-
-    Response::builder().body(response)
+    })
+    .into_response()
 }
 
 #[handler]
@@ -279,76 +265,11 @@ async fn return_auth_details(Path(_identifier): Path<String>) -> impl IntoRespon
     todo!()
 }
 
-struct TokenMiddleware {
-    token: String,
-}
-
-impl<E: Endpoint> Middleware<E> for TokenMiddleware {
-    type Output = TokenMiddlewareImpl<E>;
-
-    fn transform(&self, ep: E) -> Self::Output {
-        TokenMiddlewareImpl {
-            ep,
-            token: self.token.clone(),
-        }
-    }
-}
-
-/// The new endpoint type generated by the TokenMiddleware.
-struct TokenMiddlewareImpl<E> {
-    ep: E,
-    token: String,
-}
-
-/// Token data
-impl<E: Endpoint> Endpoint for TokenMiddlewareImpl<E> {
-    type Output = E::Output;
-
-    async fn call(&self, req: Request) -> poem::Result<Self::Output> {
-        if req.method() == Method::OPTIONS {
-            return self.ep.call(req).await;
-        }
-
-        if let Some(value) = req
-            .headers()
-            .get("Authorization")
-            .and_then(|value| value.to_str().ok())
-        {
-            let token = value.to_string();
-
-            if token == format!("Bearer {}", self.token) {
-                self.ep.call(req).await
-            } else {
-                Err(poem::Error::from_string(
-                    "Token does not match",
-                    StatusCode::UNAUTHORIZED,
-                ))
-            }
-        } else {
-            Err(poem::Error::from_string(
-                "No token provided",
-                StatusCode::UNAUTHORIZED,
-            ))
-        }
-    }
-}
-
-#[handler]
-async fn system_query_options() -> impl IntoResponse {
-    Response::builder().status(StatusCode::NO_CONTENT).finish()
-}
-
 pub async fn serve(config: AlerionConfig) {
     let system_endpoint = get(process_system_query)
-        .options(system_query_options)
-        .with(
-            Cors::new()
-                .allow_methods(["GET", "OPTIONS"])
-                .allow_credentials(true),
-        )
-        .with(TokenMiddleware {
-            token: config.auth.token,
-        });
+        .options(endpoint::make_sync(|_| StatusCode::NO_CONTENT))
+        .with(Cors::new().allow_credentials(true))
+        .with(BearerAuthMiddleware::new(config.auth.token));
 
     let api = Route::new().nest(
         "api",
@@ -362,3 +283,5 @@ pub async fn serve(config: AlerionConfig) {
         .run(api)
         .await;
 }
+
+pub mod middleware;
