+++ crates/alerion_core/src/webserver/middleware.rs
@@ -1,0 +1,1 @@
+pub mod bearer_auth;
